Refactor audio attack head handling for improved performance
- Update attack head data structure to use int16 format, allowing for 256 heads of up to 512 samples each, enhancing memory efficiency.
- Modify related functions and documentation to reflect changes in attack head upload commands and voice assignment, ensuring compatibility with the new structure.
- Improve command usage clarity in protocol documentation, including updates to the CDC upload process and voice assignment commands.

These changes enhance the audio processing capabilities of the Channel Card, providing better performance and sound quality.

+++ protocol/include/cardlink/audio/wave_loader.hpp
@@ -16,7 +16,7 @@ namespace cardlink {
 namespace audio {
 
 struct LoadedWave {
-  std::vector<int32_t> attack; /**< Q31 @ 48 kHz, length ≤ kAttackSamples */
+  std::vector<int16_t> attack; /**< Q15 @ 48 kHz, length ≤ kAttackSamples */
   std::vector<int16_t> body;   /**< from (len-overlap), 48 kHz int16 */
   uint32_t src_rate_hz = kSampleRateHz;
 };
